refactor vite/wxt plugin
