+++ wxt.config.ts
@@ -3,7 +3,7 @@ import tailwindcss from "@tailwindcss/vite";
 import preact from "@preact/preset-vite";
 import svgr from 'vite-plugin-svgr'
 import { version } from './package.json';
-import { injectOauthEnv } from './vite.config';
+import { injectOauthEnv } from './inject-oauth-plugin';
 import type { PluginOption } from 'vite';
 
 // See https://wxt.dev/api/config.html
